+++ index.html
@@ -461,6 +461,7 @@
         Загружаем...
       </div>
     </template>
-    <script src="js/main.js"></script>
+
+    <script src="js/main.js" type="module"></script>
   </body>
 </html>
