+++ js/data-arrays.js
@@ -1,0 +1,28 @@
+const COMMENTS_NAMES = ['Мирон','Юлия','Никита','Ксения','Роман','Марьям'];
+
+const AVATAR_IDS = [2,4,1,6,3,5];
+
+const descriptionLists = [
+  'Виды на..',
+  'Хотел бы я сюда еще вернуться',
+  'Полет нормальный',
+  'Ну как то так',
+  'Как вам такой момент ?',
+  'Только не говорите, что горизонт завален',
+  'Ммм..',
+  'Что скажете ?',
+  'Как сейчас помню те эмоции',
+  'Отдыхаемс =)',
+  'Невероятно, что тут еще добавить',
+  'Правила третей, во всей красе'];
+
+const commentsLists = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {COMMENTS_NAMES,AVATAR_IDS,descriptionLists,commentsLists};
