+++ js/data.js
@@ -1,0 +1,11 @@
+import {getLengthString,getPhotosData} from './util.js';
+
+const SIMILAR_ARRAY_COUNT = 26
+
+const COMMENT_LENGTH = 140;
+
+const commentString = 'Функция для проверки максимальной длины строки';
+
+getLengthString(commentString, COMMENT_LENGTH);
+
+getPhotosData(SIMILAR_ARRAY_COUNT);
